Add structural code

+++ app/src/main/java/io/johnliu/elementtd/gamestate/State.java
@@ -7,7 +7,7 @@ import android.view.ScaleGestureDetector;
 public abstract class State {
 
     public abstract void update();
-    public abstract void render(Canvas canvas, long deltaTime);
+    public abstract void render(Canvas canvas, float deltaTime);
 
     public void onTap(MotionEvent e) {}
     public void onScroll(MotionEvent e1, MotionEvent e2, float x, float y) {}
